pyproject.toml
[project]
name = "cv_matching_task"
version = "0.1.0"
description = "CV matching task application"
authors = [
    {name = "Nikhil Maheshwari", email = "nikhilmaheshwariii10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pypdf2>=3.0.1,<4.0.0",
    "psycopg2>=2.9.10,<3.0.0", 
    "google-generativeai>=0.8.5,<0.9.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "chromadb>=1.0.20,<2.0.0",
    "sentence-transformers>=5.1.0,<6.0.0",
    "litellm>=1.76.0,<2.0.0",
    "pdf2image>=1.17.0,<2.0.0",
    "scikit-learn>=1.7.1,<2.0.0",
    "fuzzywuzzy>=0.18.0,<0.19.0",
    "numpy>=2.3.2,<3.0.0",
    "tabulate>=0.9.0,<0.10.0",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "streamlit (>=1.49.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
